+++ requirements.txt
@@ -1,0 +1,4 @@
+requests==2.11.1
+six==1.10.0
+slackclient==1.0.1
+websocket-client==0.37.0
