+++ run.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+from slackbot import core
+
+if __name__ == "__main__":
+    core.start()
